libnpet/requirements.txt

numpy
scipy
pyvista
open3d
biopython
scikit-learn
pydantic>=2
requests
plyfile